fix: implement smooth calendar/list view transitions with DOM optimization
- Replace complex animation logic with ViewTransitionContainer for stable DOM
- Fix ViewToggle button "wackling" by using consistent font-semibold
- Optimize slider animation with debounced resize handling
- Extract BookingCard logic into separate BookingListView component
- Remove height constraints from CompactBookingList for better layout
- Implement proper fade-out -> content-switch -> fade-in sequence
- Fix ESLint warnings for unused imports in test files
