+++ src/frontend/app/components/__tests__/CalendarViewSkeleton.test.tsx
@@ -1,4 +1,4 @@
-import { render, screen } from '@testing-library/react';
+import { render } from '@testing-library/react';
 import '@testing-library/jest-dom';
 import CalendarViewSkeleton from '../CalendarViewSkeleton';
