+++ src/frontend/app/components/__tests__/CompactBookingListSkeleton.test.tsx
@@ -1,4 +1,4 @@
-import { render, screen } from '@testing-library/react';
+import { render } from '@testing-library/react';
 import '@testing-library/jest-dom';
 import CompactBookingListSkeleton from '../CompactBookingListSkeleton';
